WIP: Protobuf encodings for all message types
!! Important !!
All changes in `discv5/encoding` in this commit need to be reverted as they are only used for the test suite that generates the correct encodings.

Add Protobuf encodings for PingMessage and FindNodeFastMessage.

When new protobuf encodings are generated for other message types, you may need to copy over the `encoded` value outputs from the test suite in to their respective `encodedPacket` variables in the “real” tests.